# docs/swagger.yaml
basePath: /
definitions:
  email-service_internal_model.Account:
    description: Holds credentials and configuration for an external email service.
    properties:
      access_key_id:
        type: string
      created_at:
        type: string
      daily_send_limit:
        type: integer
      domain:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  email-service_internal_model.AccountSender:
    description: Represents the many-to-many relationship between accounts and senders,
      with specific sending configurations.
    properties:
      account:
        $ref: '#/definitions/email-service_internal_model.Account'
      account_id:
        type: integer
      created_at:
        type: string
      daily_send_limit:
        type: integer
      email_address:
        type: string
      id:
        type: integer
      sender:
        $ref: '#/definitions/email-service_internal_model.Sender'
      sender_id:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      weight:
        type: integer
    type: object
  email-service_internal_model.EmailSendRecord:
    description: Logs the details and delivery status of each individual email.
    properties:
      account_sender:
        $ref: '#/definitions/email-service_internal_model.AccountSender'
      account_sender_id:
        type: integer
      aliyun_task_id:
        type: string
      body:
        type: string
      error_message:
        type: string
      id:
        type: integer
      last_status_update_at:
        type: string
      recipient_email:
        type: string
      sent_at:
        type: string
      status:
        type: string
      subject:
        type: string
      task:
        $ref: '#/definitions/email-service_internal_model.EmailTask'
      task_id:
        type: integer
    type: object
  email-service_internal_model.EmailTask:
    description: Defines a task to send a specific email (or template) to a list of
      recipients.
    properties:
      account_sender:
        $ref: '#/definitions/email-service_internal_model.AccountSender'
      account_sender_id:
        type: integer
      aliyun_tag_id:
        type: string
      aliyun_tag_name:
        description: Fields for tracking via Aliyun API
        type: string
      body:
        type: string
      clickRate:
        type: number
      created_at:
        type: string
      created_by_user:
        $ref: '#/definitions/email-service_internal_model.User'
      created_by_user_id:
        type: integer
      id:
        type: integer
      openRate:
        type: number
      priority:
        type: integer
      recipient_group:
        $ref: '#/definitions/email-service_internal_model.RecipientGroup'
      recipient_group_id:
        type: integer
      scheduled_at:
        type: string
      status:
        type: string
      subject:
        type: string
      task_name:
        type: string
      template:
        $ref: '#/definitions/email-service_internal_model.EmailTemplate'
      template_id:
        type: integer
      total_click_count:
        type: integer
      total_open_count:
        type: integer
      unique_click_count:
        type: integer
      unique_open_count:
        type: integer
      uniqueClickRate:
        type: number
      uniqueOpenRate:
        type: number
      updated_at:
        type: string
    type: object
  email-service_internal_model.EmailTemplate:
    description: Stores predefined subject and body for emails.
    properties:
      body:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      subject:
        type: string
      updated_at:
        type: string
    type: object
  email-service_internal_model.Recipient:
    description: Stores information about a person who can receive emails, including
      custom data.
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      metadata:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
      updated_at:
        type: string
    type: object
  email-service_internal_model.RecipientGroup:
    description: Defines a group of recipients, which can be a fixed list (static)
      or rule-based (dynamic).
    properties:
      created_at:
        type: string
      created_by_user:
        $ref: '#/definitions/email-service_internal_model.User'
      created_by_user_id:
        type: integer
      description:
        type: string
      group_type:
        description: '''static'' or ''dynamic'''
        type: string
      id:
        type: integer
      members:
        items:
          $ref: '#/definitions/email-service_internal_model.Recipient'
        type: array
      name:
        type: string
      rules:
        items:
          $ref: '#/definitions/email-service_internal_model.RecipientGroupRule'
        type: array
      updated_at:
        type: string
    type: object
  email-service_internal_model.RecipientGroupRule:
    description: A rule used to dynamically include recipients in a group based on
      their attributes.
    properties:
      field:
        description: e.g., "status", "metadata.country"
        type: string
      group_id:
        type: integer
      id:
        type: integer
      operator:
        description: e.g., "equals", "contains"
        type: string
      value:
        type: string
    type: object
  email-service_internal_model.Sender:
    description: Contains information about the sender of an email.
    properties:
      contact_info:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  email-service_internal_model.User:
    description: Defines a user of this email service platform, with roles and access
      rights.
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_admin:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    type: object
  email-service_internal_repository.DailyStatistics:
    properties:
      click_count:
        type: integer
      click_rate:
        type: number
      date:
        type: string
      open_count:
        type: integer
      open_rate:
        type: number
      sent_count:
        type: integer
      unique_click_count:
        type: integer
      unique_click_rate:
        type: number
      unique_open_count:
        type: integer
      unique_open_rate:
        type: number
    type: object
  email-service_internal_repository.OverallStatistics:
    properties:
      overall_click_rate:
        type: number
      overall_open_rate:
        type: number
      overall_unique_click_rate:
        type: number
      overall_unique_open_rate:
        type: number
      total_clicked:
        type: integer
      total_opened:
        type: integer
      total_sent:
        type: integer
      total_unique_clicked:
        type: integer
      total_unique_opened:
        type: integer
    type: object
  email-service_internal_repository.SenderStatistics:
    properties:
      account_name:
        type: string
      account_sender_id:
        type: integer
      click_count:
        type: integer
      click_rate:
        type: number
      open_count:
        type: integer
      open_rate:
        type: number
      sender_email:
        type: string
      sender_name:
        type: string
      sent_count:
        type: integer
      unique_click_count:
        type: integer
      unique_click_rate:
        type: number
      unique_open_count:
        type: integer
      unique_open_rate:
        type: number
    type: object
  email-service_internal_service.PreviewData:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      metadata:
        additionalProperties: true
        type: object
    type: object
  email-service_internal_service.RenderedTemplate:
    properties:
      body:
        type: string
      subject:
        type: string
    type: object
  email-service_internal_service.StatisticsPeriod:
    properties:
      days:
        type: integer
      end_date:
        type: string
      start_date:
        type: string
    type: object
  email-service_internal_service.StatisticsResponse:
    properties:
      by_sender:
        items:
          $ref: '#/definitions/email-service_internal_repository.SenderStatistics'
        type: array
      period:
        $ref: '#/definitions/email-service_internal_service.StatisticsPeriod'
      summary:
        $ref: '#/definitions/email-service_internal_repository.OverallStatistics'
      time_series:
        items:
          $ref: '#/definitions/email-service_internal_repository.DailyStatistics'
        type: array
    type: object
  email-service_internal_service.TaskSummary:
    description: Provides a high-level overview of an email task's performance.
    properties:
      click_count:
        example: 250
        type: integer
      click_rate:
        example: 0.16
        type: number
      created_at:
        example: "2023-10-26 10:00:00"
        type: string
      id:
        example: 1
        type: integer
      open_count:
        example: 800
        type: integer
      open_rate:
        example: 0.53
        type: number
      sender_email:
        example: marketing@example.com
        type: string
      status:
        example: completed
        type: string
      task_name:
        example: Q3 Newsletter
        type: string
      total_recipients:
        example: 1500
        type: integer
      unique_click_count:
        example: 200
        type: integer
      unique_click_rate:
        example: 0.13
        type: number
      unique_open_count:
        example: 600
        type: integer
      unique_open_rate:
        example: 0.4
        type: number
    type: object
...
  internal_handler.CreateTemplateRequest:
    description: Request body for creating a new email template.
    properties:
      body:
        example: <h1>Hello {{.FirstName}}!</h1><p>Thank you for joining us.</p>
        type: string
      name:
        example: Welcome Email
        type: string
      subject:
        example: Welcome to Our Service!
        type: string
    required:
    - body
    - name
    - subject
    type: object
...
  internal_handler.Response:
    properties:
      data: {}
      error:
        type: string
    type: object
...
  internal_handler.UpdateTemplateRequest:
    description: Request body for updating an existing email template.
    properties:
      body:
        example: <h2>Hello {{.FirstName}}!</h2><p>We are thrilled to have you.</p>
        type: string
      name:
        example: Welcome Email V2
        type: string
      subject:
        example: A Warm Welcome to Our Service!
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: email-service/internal/handler
    url: http://www.swagger.io/support
  description: This is a powerful email service API for batch sending and management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Email Service API
  version: "1.0"
paths:
...
  /api/v1/statistics:
    get:
      description: |-
        Get comprehensive email sending statistics with various filters and grouping options.
        Provides overall summary, time series data, and sender performance analysis.
      parameters:
      - description: Start date (YYYY-MM-DD format)
        example: "2023-10-01"
        in: query
        name: start_date
        type: string
      - description: End date (YYYY-MM-DD format)
        example: "2023-10-31"
        in: query
        name: end_date
        type: string
      - description: Filter by specific account ID
        example: 1
        in: query
        name: account_id
        type: integer
      - description: Filter by specific sender ID
        example: 1
        in: query
        name: account_sender_id
        type: integer
      - description: 'Group results by: day, week, month, sender'
        enum:
        - day
        - week
        - month
        - sender
        example: day
        in: query
        name: group_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The data field contains the comprehensive statistics
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_service.StatisticsResponse'
              type: object
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Multi-Dimensional Statistics
      tags:
      - Statistics
...
  /api/v1/tasks/{id}:
    get:
      description: Get a summary of sending statistics for a specific task.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: The data field will contain the task summary
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_service.TaskSummary'
              type: object
        "400":
          description: Invalid Task ID
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Task Summary
      tags:
      - Statistics
  /api/v1/tasks/{id}/records:
    get:
      description: Get a detailed list of all email send records for a specific task.
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/email-service_internal_model.EmailSendRecord'
                  type: array
              type: object
        "400":
          description: Invalid Task ID
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Task not found or no records for this task
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Task Records
      tags:
      - Statistics
  /api/v1/templates:
    get:
      description: Retrieves a list of all email templates.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/email-service_internal_model.EmailTemplate'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get All Templates
      tags:
      - Templates
    post:
      consumes:
      - application/json
      description: Creates a new email template.
      parameters:
      - description: Template Details
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateTemplateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.EmailTemplate'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Create Template
      tags:
      - Templates
  /api/v1/templates/{id}:
    delete:
      description: Deletes an email template by its ID.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Template deleted successfully'
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Invalid template ID
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete Template
      tags:
      - Templates
    get:
      description: Retrieves a specific email template by its ID.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.EmailTemplate'
              type: object
        "400":
          description: Invalid template ID
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Get Template by ID
      tags:
      - Templates
    put:
      consumes:
      - application/json
      description: Updates an existing email template.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Template Details
        in: body
        name: template
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateTemplateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_model.EmailTemplate'
              type: object
        "400":
          description: Invalid template ID or request body
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Update Template
      tags:
      - Templates
  /api/v1/templates/{id}/preview:
    post:
      consumes:
      - application/json
      description: Renders a preview of a template with sample data.
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: integer
      - description: Sample data for preview
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/email-service_internal_service.PreviewData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/internal_handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/email-service_internal_service.RenderedTemplate'
              type: object
        "400":
          description: Invalid template ID or request body
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "404":
          description: Template not found
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - ApiKeyAuth: []
      summary: Preview Template
      tags:
      - Templates
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"